+++ hardhat.config.js
@@ -9,6 +9,7 @@ const config = {
     version: "0.8.20",
     settings: {
       optimizer: { enabled: true, runs: 200 },
+      viaIR: true,
     },
   },
   paths: {